Break off functional elements of train into their own functions.
Two subfunctions were identified within the train method and broken out to their own functions in order to clarify aand highlight the scripts workflow. The save_model function was created, this function simply saves the various components of a model to the results dictionary. The select_model function was created, this function takes an inputted model name and tries to match it with one of the existing models. If it finds a match it returns the names of all requisite files required to interact with the model class and its parameters; this include the name of its config file, the name of it's associated module, and the name of the class which implements the model within the module. Select_model retrieves this information from the config_file of the chosen model, which now contains the new section 'Aliases'.