rename network components, supervise generator output
The network component variables were all renamed to give an intuitive intuition of how each component generates its corresponding output. This was done by using the following convention of labeling the output of any network by prepending the input symbol (x for real data, z for noise) with the tags of the network components the input passed through to generate te output. For example, the autoencoder generates its output by passing real data through and embedder and then through a recovery function, as such its output is labeled re_x indicating the function that was used r(e(x)). The tags for eachcomponent are as follows: recoverer (r), embedder (e), generator (g), discriminator (d), supervisor (s).

The next change made in this commit is that the supervised loss is now defined as the mse of the outputs sg_z with e_x, opposed tose_x with e_x (which isn't really a supervised loss, as s_x is simply coerced to be the identity function).