Allow for the dynamic evaluation of a saved model.
This commit allows the user to specify functions from the evaluate module in a saved modules corresponding config file, to be run on the model itself. Here we introduce and ensure three guidelines are respected throughout the module: Firstly, it is ensured that each function in the module is 'stand-alone' that meaning that each function relies solely on the parameters of the chosen model, and the model itself to generate its output. Secondly, each function works on batch_level data, with this we avoid any ambiguity over the dimension of data passed to any of the evaluate functions. There do exist protected functions that don't work on batch level data, for this reason they remain protected. Finally, each function in the evaluate module expects an instance of a models associated class, not the raw keras object, this is to ensure that all of the parameters tied to any instance of the model remain unambiguously accessible to the function.