Split train to two methods.
The train method was split to retrieve_params and instantiate model. Before this commit parameters parsed in train were lost upon method completion. Furthermore the name 'train' incorrectly implied that this module provided the user with the ability to train the model; this is is provided by a given model class' own train method. Now the function of this module is clarified, and its results persist for later use.

A more insignificant change was the displacement of the code run when run as '__main__' to the function main.