Integrate the modified train module into evaluate.
The functions instantiate_model and retrieve_params added in a previous commit on the train branch and pulled here from main were utilized to simplify the code. Furthermore, a bug was fixed in the evaluate_model method; two objects entitled generator were getting passed to the sample function, ones through kwargs and once through manually passing the generator to the function.