Add the ability to cache evaluations in hist.
The user now has the option to designate the new cache argument as true or false in the mdoels config. If they designate it as truethen the user should also designate a name for the evaluation, now present in the 'Aliases' section of the config. The evaluation carried out will then be saved in the hist folder in a subdirectory for the given day of the evaluation, with the chosen evaluation name alias. It is now enforced that the evaluation dir only contains the most recent evaluation, if an evaluation is not cached, and a new one is carried out, it will be lost